Worked on #711 (Support for .NET 2.0) ClassDefCol - Removed unneccessary use of Linq in FindByClassName RecordingExceptionNotifier - Removed unneccessary use of Linq XmlClassLoader - Removed unneccessary use of Linq RelationshipCol - Removed unneccessary use of Linq
Fixed #566 (BOCol.LoadWithLimit with null string order criteria resulting in error)
Fixed #565 (BOCol.LoadWithLimit with null string criteria resulting in error)
Fixed #567 (LoadWithLimit with Composite Sort Criteria results in invalid SQL)

+++ source/Habanero.Test/ContactPersonCompositeKey.cs
@@ -187,10 +187,26 @@ namespace Habanero.Test
 
         public static void DeleteAllContactPeople()
         {
+            if (DatabaseConnection.CurrentConnection == null) return;
             const string sql = "DELETE FROM ContactPersonCompositeKey";
             DatabaseConnection.CurrentConnection.ExecuteRawSql(sql);
         }
 
+        public static ContactPersonCompositeKey CreateSavedContactPerson()
+        {
+            ContactPersonCompositeKey contact = CreateUnsavedContactPerson();
+
+            contact.Save();
+            return contact;
+        }
+
+        public static ContactPersonCompositeKey CreateUnsavedContactPerson()
+        {
+            ContactPersonCompositeKey contact = new ContactPersonCompositeKey();
+            contact.PK1Prop1 = TestUtil.GetRandomString();
+            contact.PK1Prop2 = TestUtil.GetRandomString();
+            return contact;
+        }
         #endregion //ForTesting
 
         public override string ToString()
